refactor(web): split login page responsibilities
