+++ apps/web/src/features/auth/OAuthSection.tsx
@@ -1,0 +1,52 @@
+import { getOAuthUrl } from '@/features/auth/oauth-urls';
+
+export function OAuthSection() {
+  const handleKakaoLogin = () => {
+    window.location.href = getOAuthUrl('kakao');
+  };
+
+  const handleGoogleLogin = () => {
+    window.location.href = getOAuthUrl('google');
+  };
+
+  return (
+    <div className="flex flex-col gap-3">
+      <button
+        type="button"
+        onClick={handleKakaoLogin}
+        className="flex min-h-11 w-full items-center justify-center gap-2 rounded-lg bg-[#FEE500] px-4 py-3 text-sm font-semibold text-[#191919] transition-opacity hover:opacity-90 focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-[var(--mmp-color-primary)] focus-visible:ring-offset-2 focus-visible:ring-offset-[var(--mmp-color-canvas)]"
+      >
+        <svg className="h-5 w-5" viewBox="0 0 24 24" fill="currentColor" aria-hidden="true">
+          <path d="M12 3C6.477 3 2 6.463 2 10.691c0 2.724 1.8 5.117 4.508 6.473-.197.735-.714 2.666-.818 3.08-.128.507.186.5.39.364.16-.106 2.55-1.734 3.58-2.44.77.112 1.562.17 2.34.17 5.523 0 10-3.463 10-7.647C22 6.463 17.523 3 12 3" />
+        </svg>
+        카카오로 시작하기
+      </button>
+
+      <button
+        type="button"
+        onClick={handleGoogleLogin}
+        className="flex min-h-11 w-full items-center justify-center gap-2 rounded-lg border border-[var(--mmp-color-hairline)] bg-white px-4 py-3 text-sm font-semibold text-gray-900 transition-opacity hover:opacity-90 focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-[var(--mmp-color-primary)] focus-visible:ring-offset-2 focus-visible:ring-offset-[var(--mmp-color-canvas)]"
+      >
+        <svg className="h-5 w-5" viewBox="0 0 24 24" aria-hidden="true">
+          <path
+            d="M22.56 12.25c0-.78-.07-1.53-.2-2.25H12v4.26h5.92a5.06 5.06 0 0 1-2.2 3.32v2.77h3.57c2.08-1.92 3.28-4.74 3.28-8.1z"
+            fill="#4285F4"
+          />
+          <path
+            d="M12 23c2.97 0 5.46-.98 7.28-2.66l-3.57-2.77c-.98.66-2.23 1.06-3.71 1.06-2.86 0-5.29-1.93-6.16-4.53H2.18v2.84C3.99 20.53 7.7 23 12 23z"
+            fill="#34A853"
+          />
+          <path
+            d="M5.84 14.09c-.22-.66-.35-1.36-.35-2.09s.13-1.43.35-2.09V7.07H2.18C1.43 8.55 1 10.22 1 12s.43 3.45 1.18 4.93l2.85-2.22.81-.62z"
+            fill="#FBBC05"
+          />
+          <path
+            d="M12 5.38c1.62 0 3.06.56 4.21 1.64l3.15-3.15C17.45 2.09 14.97 1 12 1 7.7 1 3.99 3.47 2.18 7.07l3.66 2.84c.87-2.6 3.3-4.53 6.16-4.53z"
+            fill="#EA4335"
+          />
+        </svg>
+        Google로 시작하기
+      </button>
+    </div>
+  );
+}
